Gen AI/requirements.txt
langchain
langchain-chroma
langchain-mistralai
chromadb
python-dotenv